--- .classpath
@@ -1,9 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<classpath>
-	<classpathentry kind="src" path="src"/>
-	<classpathentry kind="con" path="org.eclipse.jdt.launching.JRE_CONTAINER"/>
-	<classpathentry kind="con" path="org.eclipse.jdt.USER_LIBRARY/Spigot1.8.8"/>
-	<classpathentry combineaccessrules="false" kind="src" path="/BridgingAnalyzer"/>
-	<classpathentry exported="true" kind="lib" path="libs/gson-2.8.5-repackage.jar"/>
-	<classpathentry kind="output" path="bin"/>
-</classpath>
