+++ .gitignore
@@ -1,2 +1,2 @@
-/.settings/
-/bin/
+.idea
+target
