+++ src/main/java/sakura/kooi/BridgingSkin/ClearThread.java
