+++ src/main/resources/skins.json
@@ -1,0 +1,1 @@
+[]
