--- src/plugin.yml
@@ -1,10 +1,0 @@
-name: BridgingSkin
-main: sakura.kooi.BridgingSkin.BridgingSkin
-version: 3
-author: SakuraKooi
-depend: ["BridgingAnalyzer"]
-commands:
-  bskin:
-    description: Choose your skin.
-    usage: /bskin
-  bskin-edit:
